python/3DRecon/requirements.txt
numpy>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
pytorch3d>=0.7.4
scipy>=1.9.0
matplotlib>=3.5.0
tqdm>=4.64.0
pycolmap>=0.3.0
open3d>=0.17.0
plyfile>=0.7.4
pyyaml>=6.0
pillow>=9.4.0
imageio>=2.25.0
trimesh>=3.9.0